# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAP Planner - Full API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
paths:
...
  /wo:
    get:
      summary: List work orders
      tags: [WorkOrders]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Work orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkOrder"
  /wo/{wo_id}:
    get:
      summary: Get work order by id or business wo_id
      tags: [WorkOrders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: wo_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Work order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    WorkOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wo_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
...